Make xorrisofs argument not prevent efi boot
"In "imgcreate/live.py", the order of options to xorrisofs produces an image that is only bootable by bios. Booting via efi fails.

Additionally, the image will not be bootable at all if efiboot.img for some reason is missing. This file should only be needed for efi, but all El-Torito options are skipped if efiboot.img is missing."

From: https://github.com/livecd-tools/livecd-tools/issues/190
Signed-off-by: Mikhail Novosyolov <m.novosyolov@rosalinux.ru>